Update Ext_T_AutoFireHelper.uc
Woo! Confirmed that the make.sh script works as intended, both for compiling and for testing. Above compiled successfully, allowed survivalist to autofire with 9mm / MP7 / AR-15 / SG-500 / HX25 / GS Start revolver / Spitfire, allowed GS to autofire with GS starter revolver + spitfire but not with MP7. Should be fully functional like this.

Thank you for the guidance! As an aside - Do you have any advice for finding out the names of variables like AssociatedPerkClass, or just 'look around until you find it in an appropriate class'?